+++ projects/ngx-mask-lib/src/lib/ngx-mask.config.ts
@@ -5,7 +5,7 @@ export type InputTransformFn = (value: unknown) => string | number;
 
 export type OutputTransformFn = (value: string | number | undefined | null) => unknown;
 
-export interface IConfig {
+export type NgxMaskConfig = {
     suffix: string;
     prefix: string;
     thousandSeparator: string;
@@ -17,7 +17,7 @@ export interface IConfig {
     shownMaskExpression: string;
     specialCharacters: string[] | readonly string[];
     dropSpecialCharacters: boolean | string[] | readonly string[];
-    hiddenInput: boolean | undefined;
+    hiddenInput: boolean;
     validation: boolean;
     separatorLimit: string;
     apm: boolean;
@@ -37,14 +37,14 @@ export interface IConfig {
             symbol?: string;
         }
     >;
-}
+};
 
-export type optionsConfig = Partial<IConfig>;
-export const NGX_MASK_CONFIG = new InjectionToken<IConfig>('ngx-mask config');
-export const NEW_CONFIG = new InjectionToken<IConfig>('new ngx-mask config');
-export const INITIAL_CONFIG = new InjectionToken<IConfig>('initial ngx-mask config');
+export type NgxMaskOptions = Partial<NgxMaskConfig>;
+export const NGX_MASK_CONFIG = new InjectionToken<NgxMaskConfig>('ngx-mask config');
+export const NEW_CONFIG = new InjectionToken<NgxMaskConfig>('new ngx-mask config');
+export const INITIAL_CONFIG = new InjectionToken<NgxMaskConfig>('initial ngx-mask config');
 
-export const initialConfig: IConfig = {
+export const initialConfig: NgxMaskConfig = {
     suffix: '',
     prefix: '',
     thousandSeparator: ' ',
@@ -54,12 +54,11 @@ export const initialConfig: IConfig = {
     showMaskTyped: false,
     placeHolderCharacter: '_',
     dropSpecialCharacters: true,
-    hiddenInput: undefined,
+    hiddenInput: false,
     shownMaskExpression: '',
     separatorLimit: '',
     allowNegativeNumbers: false,
     validation: true,
-
     specialCharacters: ['-', '/', '(', ')', '.', ':', ' ', '+', ',', '@', '[', ']', '"', "'"],
     leadZeroDateTime: false,
     apm: false,