+++ projects/ngx-mask-lib/src/test/custom-patterns.spec.ts
@@ -1,11 +1,13 @@
-import { ComponentFixture, TestBed } from '@angular/core/testing';
+import type { ComponentFixture } from '@angular/core/testing';
+import { TestBed } from '@angular/core/testing';
 import { ReactiveFormsModule } from '@angular/forms';
 
 import { TestMaskComponent } from './utils/test-component.component';
 import { equal } from './utils/test-functions.component';
 import { provideNgxMask } from '../lib/ngx-mask.providers';
 import { NgxMaskDirective } from '../lib/ngx-mask.directive';
-import { IConfig, initialConfig } from 'ngx-mask';
+import type { NgxMaskConfig } from 'ngx-mask';
+import { initialConfig } from 'ngx-mask';
 
 describe('Directive: Mask (Custom patterns)', () => {
     let fixture: ComponentFixture<TestMaskComponent>;
@@ -42,7 +44,7 @@ describe('Directive: Mask (Provide custom patterns)', () => {
         B: { pattern: new RegExp('[a-zA-Z]'), optional: true },
     };
 
-    const ngxMaskConfigValue: Partial<IConfig> = {
+    const ngxMaskConfigValue: Partial<NgxMaskConfig> = {
         patterns: { ...initialConfig.patterns, ...ngxMaskCustomPatterns },
     };
 
@@ -86,7 +88,7 @@ describe('Directive: Mask (Provide custom patterns with symbol *)', () => {
         '*': { pattern: new RegExp('[a-zA-Z0-9]') },
     };
 
-    const ngxMaskConfigValue: Partial<IConfig> = {
+    const ngxMaskConfigValue: Partial<NgxMaskConfig> = {
         patterns: { ...initialConfig.patterns, ...ngxMaskCustomPatterns },
     };
 
@@ -133,7 +135,7 @@ describe('Directive: Mask (Provide custom patterns with symbol f and F)', () => 
     let fixture: ComponentFixture<TestMaskComponent>;
     let component: TestMaskComponent;
 
-    const ngxMaskConfig: IConfig = {
+    const ngxMaskConfig: NgxMaskConfig = {
         ...initialConfig,
         patterns: {
             f: {
@@ -188,7 +190,7 @@ describe('Directive: Mask (Provide custom patterns with symbol B optional)', () 
         B: { pattern: new RegExp('[0-9,;]'), optional: true },
     };
 
-    const ngxMaskConfigValue: Partial<IConfig> = {
+    const ngxMaskConfigValue: Partial<NgxMaskConfig> = {
         patterns: { ...initialConfig.patterns, ...ngxMaskCustomPatterns },
     };