+++ projects/ngx-mask-lib/src/test/separator.spec.ts
@@ -1,6 +1,7 @@
-import { ComponentFixture, fakeAsync, TestBed, tick } from '@angular/core/testing';
+import type { ComponentFixture } from '@angular/core/testing';
+import { fakeAsync, TestBed, tick } from '@angular/core/testing';
 import { By } from '@angular/platform-browser';
-import { DebugElement } from '@angular/core';
+import type { DebugElement } from '@angular/core';
 import { ReactiveFormsModule } from '@angular/forms';
 import { TestMaskComponent } from './utils/test-component.component';
 import { equal, typeTest } from './utils/test-functions.component';
@@ -1724,7 +1725,6 @@ describe('Separator: Mask', () => {
         fixture.detectChanges();
 
         inputElement.setSelectionRange(2, 2);
-        inputElement.selectionStart;
         expect(inputElement.selectionStart).toBe(2);
 
         const backspaceEvent = new KeyboardEvent('keydown', {
@@ -1753,7 +1753,6 @@ describe('Separator: Mask', () => {
         fixture.detectChanges();
 
         inputElement.setSelectionRange(2, 2);
-        inputElement.selectionStart;
         expect(inputElement.selectionStart).toBe(2);
 
         const backspaceEvent = new KeyboardEvent('keydown', {
@@ -1782,7 +1781,6 @@ describe('Separator: Mask', () => {
         fixture.detectChanges();
 
         inputElement.setSelectionRange(2, 2);
-        inputElement.selectionStart;
         expect(inputElement.selectionStart).toBe(2);
 
         const backspaceEvent = new KeyboardEvent('keydown', {