+++ projects/ngx-mask-lib/src/test/utils/cypress-test-component.component.ts
@@ -1,7 +1,8 @@
 import { Component, inject, Input } from '@angular/core';
 import { FormControl } from '@angular/forms';
 import { scan, startWith } from 'rxjs';
-import { IConfig, NGX_MASK_CONFIG } from 'ngx-mask';
+import type { NgxMaskConfig } from 'ngx-mask';
+import { NGX_MASK_CONFIG } from 'ngx-mask';
 
 @Component({
     selector: 'jsdaddy-open-source-test',
@@ -33,7 +34,7 @@ import { IConfig, NGX_MASK_CONFIG } from 'ngx-mask';
     `,
 })
 export class CypressTestMaskComponent {
-    protected _config = inject<IConfig>(NGX_MASK_CONFIG);
+    protected _config = inject<NgxMaskConfig>(NGX_MASK_CONFIG);
     @Input() public mask = '';
 
     @Input() public hiddenInput = false;
@@ -71,6 +72,6 @@ export class CypressTestMaskComponent {
     public readonly counter$ = this.form.valueChanges.pipe(
         startWith(0),
 
-        scan((_, __, index) => ++index)
+        scan((acc) => acc + 1, 0)
     );
 }