+++ src/main.ts
@@ -8,9 +8,11 @@ import { provideHttpClient } from '@angular/common/http';
 import { BaseHttpService } from '@libraries/base-http/base-http.service';
 import { DOMAIN } from '@libraries/token/token';
 import { GithubStarsService } from '@libraries/github/github-stars.service';
+import { provideExperimentalZonelessChangeDetection } from '@angular/core';
 
 bootstrapApplication(AppComponent, {
     providers: [
+        provideExperimentalZonelessChangeDetection(),
         GithubStarsService,
         provideHttpClient(),
         {
@@ -28,4 +30,5 @@ bootstrapApplication(AppComponent, {
             },
         },
     ],
+    // eslint-disable-next-line no-console
 }).catch((err) => console.error(err));